chore: move signing wrapper (deprecated)
