+++ src/modules/common/components/signInTooltipWrapper/signInTooltipWrapper.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import routes from '@screens/router/routes';
-import Tooltip from 'src/theme/Tooltip';
+import Tooltip from '@theme/Tooltip';
 import styles from './signInTooltipWrapper.css';
 
 const SignInTooltipWrapper = ({
