+++ src/modules/token/fungible/components/BalanceInfo/ActionBar.js
@@ -9,7 +9,7 @@ import {
 import DialogLink from 'src/theme/dialog/link';
 import Tooltip from 'src/theme/Tooltip';
 import { selectAccountBalance, selectLSKAddress } from '@common/store/selectors';
-import SignInTooltipWrapper from '@packages/wallet/manage/signInTooltipWrapper'; // @todo move to the token domain
+import SignInTooltipWrapper from 'src/modules/common/components/signInTooltipWrapper';
 import EmptyBalanceTooltipWrapper from './EmptyBalanceTooltipWrapper';
 import styles from './BalanceInfo.css';
